requirements.txt
asgiref==3.8.1
bandit==1.7.10
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
Django==4.2.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
flake8==7.1.1
google-api-core==2.21.0
google-auth==2.35.0
google-cloud-vision==3.7.4
googleapis-common-protos==1.65.0
grpcio==1.67.0
grpcio-status==1.67.0
idna==3.10
iniconfig==2.0.0
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
packaging==24.2
pbr==6.1.0
pluggy==1.5.0
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pyflakes==3.2.0
Pygments==2.18.0
pytest==8.3.3
pytest-django==4.9.0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
rich==13.9.3
rsa==4.9
sqlparse==0.5.1
stevedore==5.3.0
tzdata==2024.2
urllib3==2.2.3
